Network abstraction layer consolidating streamer into net-utils
