+++ banking-bench/Cargo.toml
@@ -16,7 +16,7 @@ rand = "0.7.0"
 rayon = "1.5.0"
 solana-core = { path = "../core", version = "=1.7.0" }
 solana-clap-utils = { path = "../clap-utils", version = "=1.7.0" }
-solana-streamer = { path = "../streamer", version = "=1.7.0" }
+solana-net-utils = { path = "../net-utils", version = "=1.7.0" }
 solana-perf = { path = "../perf", version = "=1.7.0" }
 solana-ledger = { path = "../ledger", version = "=1.7.0" }
 solana-logger = { path = "../logger", version = "=1.7.0" }
