+++ banks-server/Cargo.toml
@@ -18,6 +18,7 @@ solana-banks-interface = { path = "../banks-interface", version = "=1.7.0" }
 solana-runtime = { path = "../runtime", version = "=1.7.0" }
 solana-sdk = { path = "../sdk", version = "=1.7.0" }
 solana-metrics = { path = "../metrics", version = "=1.7.0" }
+solana-net-utils = { path = "../net-utils", version = "=1.7.0" }
 tarpc = { version = "0.24.1", features = ["full"] }
 tokio = { version = "1", features = ["full"] }
 tokio-serde = { version = "0.8", features = ["bincode"] }
