+++ core/src/validator.rs
@@ -45,6 +45,7 @@ use solana_ledger::{
 };
 use solana_measure::measure::Measure;
 use solana_metrics::datapoint_info;
+use solana_net_utils::SocketLike;
 use solana_runtime::{
     accounts_index::AccountIndex,
     bank::Bank,
