+++ net-utils/Cargo.toml
@@ -13,6 +13,7 @@ edition = "2018"
 bincode = "1.3.1"
 clap = "2.33.1"
 log = "0.4.11"
+libc = "0.2.81"
 nix = "0.19.0"
 rand = "0.7.0"
 serde = "1.0.122"
@@ -20,9 +21,15 @@ serde_derive = "1.0.103"
 socket2 = "0.3.17"
 solana-clap-utils = { path = "../clap-utils", version = "=1.7.0" }
 solana-logger = { path = "../logger", version = "=1.7.0" }
+solana-measure = { path = "../measure", version = "=1.7.0" }
+solana-metrics = { path = "../metrics", version = "=1.7.0" }
+solana-perf = { path = "../perf", version = "=1.7.0" }
+solana-sdk = { path = "../sdk", version = "=1.7.0" }
 solana-version = { path = "../version", version = "=1.7.0" }
+thiserror = "1.0"
 tokio = { version = "1", features = ["full"] }
 url = "2.1.1"
+enum_dispatch = "0.3.5"
 
 [lib]
 name = "solana_net_utils"
