+++ net-utils/src/streamer.rs
@@ -1,0 +1,8 @@
+#![allow(clippy::integer_arithmetic)]
+pub mod streamer;
+pub mod recvmmsg;
+pub mod sendmmsg;
+pub mod packet;
+
+pub use streamer::*;
+pub use packet::*;
