--- streamer/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "solana-streamer"
-version = "1.7.0"
-description = "Solana Streamer"
-authors = ["Solana Maintainers <maintainers@solana.foundation>"]
-repository = "https://github.com/solana-labs/solana"
-license = "Apache-2.0"
-homepage = "https://solana.com/"
-documentation = "https://docs.rs/solana-streamer"
-edition = "2018"
-
-[dependencies]
-log = "0.4.11"
-solana-metrics = { path = "../metrics", version = "=1.7.0" }
-solana-sdk = { path = "../sdk", version = "=1.7.0" }
-thiserror = "1.0"
-solana-measure = { path = "../measure", version = "=1.7.0" }
-solana-logger = { path = "../logger", version = "=1.7.0" }
-libc = "0.2.81"
-nix = "0.19.0"
-solana-perf = { path = "../perf", version = "=1.7.0" }
-
-[dev-dependencies]
-
-[lib]
-crate-type = ["lib"]
-name = "solana_streamer"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
