--- streamer/src/lib.rs
@@ -1,11 +1,0 @@
-#![allow(clippy::integer_arithmetic)]
-pub mod packet;
-pub mod recvmmsg;
-pub mod sendmmsg;
-pub mod streamer;
-
-#[macro_use]
-extern crate log;
-
-#[macro_use]
-extern crate solana_metrics;
